13-the-yule-lads/functions/requirements.txt
azure-functions
tensorflow==1.14.0
textgenrnn==1.5.0